requirements.txt
absl-py==0.10.0
appdirs==1.4.4
cachetools==4.1.1
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
future==0.18.2
google-auth==1.21.2
google-auth-oauthlib==0.4.1
grpcio==1.32.0
identify==1.5.4
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
networkx==2.5
nodeenv==1.5.0
numpy==1.19.2
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
Pillow-SIMD==7.0.0.post3
pre-commit==2.7.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-dotenv==0.14.0
pytorch-lightning==0.9.0
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
tifffile==2020.9.3
toml==0.10.1
torch==1.6.0
torch==1.6.0
torchvision==0.7.0
torchvision==0.7.0
tqdm==4.49.0
tqdm==4.49.0
urllib3==1.25.10
virtualenv==20.0.31
Werkzeug==1.0.1
zipp==3.1.0
